move calib.camera package to calib.
