+++ digicampipe/calib/image.py
